resnet18/requirements.txt
# renet requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
numpy==1.23.0
Pillow==9.0.1